requirements.txt
asgiref==3.6.0
Django==4.2
django-modeltranslation==0.18.9
Pillow==9.5.0
psycopg2==2.9.6
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
